[BE] 닉네임 중복검증 API Redis 수정

+++ BE/src/main/java/com/newsainturtle/shadowmate/user/service/UserServiceImpl.java
@@ -67,11 +67,20 @@ public class UserServiceImpl implements UserService {
     @Override
     @Transactional
     public void updateUser(final Long userId, final UpdateUserRequest updateUserRequest) {
+        final User user = userRepository.findByIdAndNickname(userId, updateUserRequest.getNewNickname());
+        if(user == null) {
+            final Boolean getHashNickname = redisService.getHashNicknameData(updateUserRequest.getNewNickname());
+            if(getHashNickname == null || !getHashNickname) {
+                throw new UserException(UserErrorResult.RETRY_NICKNAME);
+            }
+            else {
+                redisService.deleteNicknameData(updateUserRequest.getNewNickname());
+            }
+        }
         userRepository.updateUser(updateUserRequest.getNewNickname(),
                 updateUserRequest.getNewProfileImage(),
                 updateUserRequest.getNewStatusMessage(),
                 userId);
-        redisService.deleteNicknameData(updateUserRequest.getNewNickname());
     }
 
     @Override
